locallibrary/requirements.txt
dj-database-url==0.5.0
Django==1.11.29
django-allauth==0.42.0
django-cors-headers==3.4.0
django-registration==2.4.1
django-registration-redux==1.6
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
lxml==4.5.1
Pillow==7.1.2
pipenv==2018.11.26
postgres==3.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
psycopg2-pool==1.1
pytils==0.3
Scrapy==1.1.0rc3
six==1.12.0
virtualenv==20.0.21
whitenoise==5.2.0

